Updating tests for global module mocker
