+++ packages/editor/src/components/Form/Fields/Select/index.tsx
@@ -1,4 +1,4 @@
-import { MenuItem, Select as MuiSelect, SelectChangeEvent, SelectProps } from '@mui/material';
+import { MenuItem, Select as MuiSelect, SelectProps } from '@mui/material';
 import styled from '@emotion/styled';
 import React from 'react';
 
@@ -120,37 +120,15 @@ export const SelectField = <Value, Option = Value>({
 }: Omit<SelectProps, 'value' | 'onChange' | 'renderValue'> & {
   value: Value;
   options: ReadonlyArray<Option>;
-  onChange: (event: { target: { value: Value } }) => void;
-  getOptionLabel: (option: Option) => React.ReactNode;
+  onChange: SelectProps<Value>['onChange'];
+  getOptionLabel: (option: Option | Value) => React.ReactNode;
   getOptionKey?: (option: Option) => string | number;
   readOnly?: boolean;
 }) => {
-  const findSelectedIndex = React.useCallback(() => {
-    // First try reference equality
-    const byRef = options.findIndex(o => o === (value as unknown as Option));
-    if (byRef !== -1) return byRef;
-    // Fallback to key comparison if provided
-    if (getOptionKey) {
-      // @ts-expect-error Value may not be Option; getOptionKey should handle compatible shape when provided
-      const targetKey = getOptionKey(value);
-      const byKey = options.findIndex(o => getOptionKey(o) === targetKey);
-      if (byKey !== -1) return byKey;
-    }
-    return -1;
-  }, [options, value, getOptionKey]);
-
-  const selectedIndex = findSelectedIndex();
-
-  const handleChange = (e: SelectChangeEvent) => {
-    const idx = Number((e.target as unknown as { value: string }).value);
-    const next = options[idx] as unknown as Value;
-    onChange({ target: { value: next } });
-  };
-
   return (
     <StyledSelect
-      onChange={handleChange as unknown as (event: unknown) => void}
-      value={selectedIndex >= 0 ? String(selectedIndex) : ''}
+      onChange={onChange}
+      value={value}
       className={`${className ?? ''} ${readOnly ? 'read-only' : ''}`}
       style={{
         ...(style as React.CSSProperties),
@@ -159,12 +137,10 @@ export const SelectField = <Value, Option = Value>({
       displayEmpty
       renderValue={selected => {
         const isEmpty = selected === '' || selected === undefined || selected === null;
-        const index = typeof selected === 'string' && selected !== '' ? parseInt(selected, 10) : selectedIndex;
-        const option = index >= 0 ? (options[index] as Option) : undefined;
-        const label = option !== undefined ? getOptionLabel(option) : '';
+        const label = getOptionLabel(selected);
         return <span style={{ color: isEmpty ? 'var(--color-text-muted)' : undefined }}>{label}</span>;
       }}
-      {...props}
+      {...(props as Partial<SelectProps<Value>>)}
     >
       {options.map((option, index) => (
         <StyledMenuItem key={getOptionKey ? getOptionKey(option) : index} value={String(index)}>
