+++ packages/editor/src/helpers/editor/createPuckComponent/createComponent.test.tsx
@@ -1,49 +1,14 @@
-import { expect, test, describe, beforeEach, afterAll, mock } from 'bun:test';
+import { expect, test, describe, beforeEach, mock } from 'bun:test';
 import { createElement } from 'react';
+import { createModuleMocker } from '@test-utils/moduleMocker';
 
 /**
  * Due to an issue with Bun (https://github.com/oven-sh/bun/issues/7823), we need to manually restore mocked modules
  * after we're done. We do this by setting the mocked value to the original module.
  */
-interface MockResult {
-  clear: () => void;
-}
-
-export class ModuleMocker {
-  private mocks: MockResult[] = [];
-
-  async mock(modulePath: string, renderMocks: () => Record<string, unknown>) {
-    const original = {
-      ...(await import(modulePath)),
-    };
-    const mocks = renderMocks();
-    const result = {
-      ...original,
-      ...mocks,
-    };
-    mock.module(modulePath, () => result);
-
-    this.mocks.push({
-      clear: () => {
-        mock.module(modulePath, () => original);
-      },
-    });
-  }
-
-  clear() {
-    this.mocks.forEach(mockResult => mockResult.clear());
-    this.mocks = [];
-  }
-}
-
-// Mock window and leaflet to prevent import issues
-(global as { window?: unknown }).window = {
-  setTimeout: setTimeout,
-  requestAnimationFrame: (fn: () => void) => setTimeout(fn, 16),
-};
 
 // Set up module mocker at top level
-const moduleMocker = new ModuleMocker();
+const moduleMocker = createModuleMocker();
 
 await moduleMocker.mock('@hakit/components', () => ({
   AvailableQueries: {},
@@ -104,11 +69,6 @@ describe('createComponent', () => {
     mockUseActiveBreakpoint.mockClear();
     mockUseGlobalStore.mockClear();
     mockUsePuckIframeElements.mockClear();
-  });
-
-  afterAll(() => {
-    // Clear all module mocks to prevent pollution of other tests
-    moduleMocker.clear();
   });
 
   const createMockComponentFactoryData = (): ComponentFactoryData => ({
